Add menu options for selecting keyboard type.
Update README to remove old dependency status badge.

Update Travis YML to have new framebuffer init.

Update gradle wrapper to support Java 11.

Make Keyboard class abstract, implement EmulatedKeyboard class.

Add passthrough keyboard functionality.

Update README to reflect new keyboard options.

+++ src/main/java/ca/craigthomas/yacoco3e/common/IO.java
@@ -7,8 +7,6 @@ package ca.craigthomas.yacoco3e.common;
 import ca.craigthomas.yacoco3e.components.Memory;
 import org.apache.commons.io.IOUtils;
 
-import javax.swing.*;
-import javax.swing.filechooser.FileFilter;
 import java.io.*;
 import java.nio.ByteOrder;
 import java.util.logging.Logger;
@@ -142,7 +140,7 @@ public class IO
         }
 
         for (byte data : source) {
-            target[byteCounter] = (short) data;
+            target[byteCounter] = data;
             byteCounter++;
         }
         return true;
